Refactored assertion helper in order to hide its visibility
